style: move unimplemented analyze modes to another file
