+++ cmd/policy-assistant/pkg/cli/analyze.go
@@ -23,22 +23,23 @@ import (
 	"github.com/spf13/cobra"
 	v1 "k8s.io/api/core/v1"
 	networkingv1 "k8s.io/api/networking/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 const (
-	ParseMode        = "parse"
-	ExplainMode      = "explain"
-	QueryTrafficMode = "query-traffic"
-	QueryTargetMode  = "query-target"
-	ProbeMode        = "probe"
+	// ParseMode        = "parse"
+	ExplainMode = "explain"
+	// QueryTrafficMode = "query-traffic"
+	// QueryTargetMode  = "query-target"
+	ProbeMode = "probe"
 )
 
+// should we remove commented out modes or implement them later?
+// code for them is in analyze_unimplemented.go
 var AllModes = []string{
-	ParseMode,
+	// ParseMode,
 	ExplainMode,
-	QueryTrafficMode,
-	QueryTargetMode,
+	// QueryTrafficMode,
+	// QueryTargetMode,
 	ProbeMode,
 }
 
@@ -158,28 +159,12 @@ func RunAnalyzeCommand(args *AnalyzeArgs) {
 	policies := matcher.BuildV1AndV2NetPols(args.SimplifyPolicies, kubePolicies, kubeANPs, kubeBANP)
 
 	for _, mode := range args.Modes {
+		// see analyze_unimplemented.go for unimplemented modes and the "case" statements for them
 		switch mode {
-		case ParseMode:
-			fmt.Println("parsed policies:")
-			ParsePolicies(kubePolicies)
 		case ExplainMode:
 			fmt.Println("explained policies:")
 			ExplainPolicies(policies)
-		case QueryTargetMode:
-			pods := make([]*QueryTargetPod, len(kubePods))
-			for i, p := range kubePods {
-				pods[i] = &QueryTargetPod{
-					Namespace: p.Namespace,
-					Labels:    p.Labels,
-				}
-			}
-			fmt.Println("query target:")
-			QueryTargets(policies, args.TargetPodPath, pods)
-		case QueryTrafficMode:
-			fmt.Println("query traffic:")
-			QueryTraffic(policies, args.TrafficPath)
 		case ProbeMode:
-			fmt.Println("probe:")
 			ProbeSyntheticConnectivity(policies, args.ProbePath, kubePods, kubeNamespaces)
 		default:
 			panic(errors.Errorf("unrecognized mode %s", mode))
@@ -187,75 +172,8 @@ func RunAnalyzeCommand(args *AnalyzeArgs) {
 	}
 }
 
-func ParsePolicies(kubePolicies []*networkingv1.NetworkPolicy) {
-	fmt.Println(kube.NetworkPoliciesToTable(kubePolicies))
-}
-
 func ExplainPolicies(explainedPolicies *matcher.Policy) {
 	fmt.Printf("%s\n", explainedPolicies.ExplainTable())
-}
-
-// QueryTargetPod matches targets; targets exist in only a single namespace and can't be matched by namespace
-//
-//	label, therefore we match by exact namespace and by pod labels.
-type QueryTargetPod struct {
-	Namespace string
-	Labels    map[string]string
-}
-
-func QueryTargets(explainedPolicies *matcher.Policy, podPath string, pods []*QueryTargetPod) {
-	if podPath != "" {
-		podsFromFile, err := json.ParseFile[[]*QueryTargetPod](podPath)
-		utils.DoOrDie(err)
-		pods = append(pods, *podsFromFile...)
-	}
-
-	for _, pod := range pods {
-		fmt.Printf("pod in ns %s with labels %+v:\n\n", pod.Namespace, pod.Labels)
-
-		targets, combinedRules := QueryTargetHelper(explainedPolicies, pod)
-
-		fmt.Printf("Matching targets:\n%s\n", targets.ExplainTable())
-		fmt.Printf("Combined rules:\n%s\n\n\n", combinedRules.ExplainTable())
-	}
-}
-
-func QueryTargetHelper(policies *matcher.Policy, pod *QueryTargetPod) (*matcher.Policy, *matcher.Policy) {
-	podInfo := &matcher.InternalPeer{
-		Namespace: pod.Namespace,
-		PodLabels: pod.Labels,
-	}
-	ingressTargets := policies.TargetsApplyingToPod(true, podInfo)
-	combinedIngressTarget := matcher.CombineTargetsIgnoringPrimaryKey(pod.Namespace, metav1.LabelSelector{MatchLabels: pod.Labels}, ingressTargets)
-
-	egressTargets := policies.TargetsApplyingToPod(false, podInfo)
-	combinedEgressTarget := matcher.CombineTargetsIgnoringPrimaryKey(pod.Namespace, metav1.LabelSelector{MatchLabels: pod.Labels}, egressTargets)
-
-	var combinedIngresses []*matcher.Target
-	if combinedIngressTarget != nil {
-		combinedIngresses = []*matcher.Target{combinedIngressTarget}
-	}
-	var combinedEgresses []*matcher.Target
-	if combinedEgressTarget != nil {
-		combinedEgresses = []*matcher.Target{combinedEgressTarget}
-	}
-
-	return matcher.NewPolicyWithTargets(ingressTargets, egressTargets), matcher.NewPolicyWithTargets(combinedIngresses, combinedEgresses)
-}
-
-func QueryTraffic(explainedPolicies *matcher.Policy, trafficPath string) {
-	if trafficPath == "" {
-		logrus.Fatalf("%+v", errors.Errorf("path to traffic file required for QueryTraffic command"))
-	}
-	allTraffics, err := json.ParseFile[[]*matcher.Traffic](trafficPath)
-	utils.DoOrDie(err)
-
-	for _, traffic := range *allTraffics {
-		fmt.Printf("Traffic:\n%s\n", traffic.Table())
-
-		result := explainedPolicies.IsTrafficAllowed(traffic)
-		fmt.Printf("Is traffic allowed?\n%s\n\n\n", result.Table())
-	}
 }
 
 type SyntheticProbeConnectivityConfig struct {
